+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- SafeFunction.getUsingDefault()
+
 ## [8.6.4] - 2024-07-18
 
 ### Fixed